test(mcp-agent): add snapshot for `github-reviewer-agent`
