+++ multimodal/mcp-agent/snapshot/github-reviewer-agent/claude/loop-2/tool-calls.jsonl
@@ -1,0 +1,17 @@
+[
+  {
+    "toolCallId": "call_1750084731916_dt0qs",
+    "name": "browser_click",
+    "args": {
+      "element": "Files changed 4",
+      "ref": "e208"
+    },
+    "result": [
+      {
+        "type": "text",
+        "text": "Error: page._wrapApiCall: Execution context was destroyed, most likely because of a navigation"
+      }
+    ],
+    "executionTime": 2139
+  }
+]
