+++ multimodal/mcp-agent/snapshot/github-reviewer-agent/volcengine/loop-1/event-stream.jsonl
@@ -1,0 +1,19 @@
+[
+  {
+    "id": "48026d12-7d17-4243-bd3e-7e4e48183032",
+    "type": "agent_run_start",
+    "timestamp": 1750077302623,
+    "sessionId": "1750077302623-nni15ax",
+    "runOptions": {
+      "input": "Review https://github.com/bytedance/UI-TARS-desktop/pull/534"
+    },
+    "provider": "volcengine",
+    "model": "ep-20250512165931-2c2ln"
+  },
+  {
+    "id": "e8e0f0f1-0f65-42cb-aff8-639a2d9d85cc",
+    "type": "user_message",
+    "timestamp": 1750077302623,
+    "content": "Review https://github.com/bytedance/UI-TARS-desktop/pull/534"
+  }
+]
